refactor: change structure of files
