+++ src/components/tree.html
@@ -98,7 +98,7 @@
       </div>
       <img class="tree-image" src="../assets/tree/2.png" usemap="#image-map" alt="tree image">
       <map class="testMap" name="image-map">
-          <area class="testMapArea" target="_blank" alt="tree" href="/check" coords="227,661,140,709,132,698,103,690,131,668,76,650,83,637,56,629,33,644,-1,638,35,613,16,585,23,580,13,551,41,559,27,525,46,521,24,505,33,482,59,472,49,451,86,422,78,412,86,404,67,376,100,356,81,322,90,318,93,296,111,288,112,271,125,267,130,256,128,229,143,222,141,187,156,187,167,132,176,112,194,81,207,55,223,33,230,33,245,1,252,1,263,9,267,36,279,44,293,61,300,75,307,96,317,96,315,126,326,124,323,165,330,161,335,190,355,188,347,214,371,220,366,230,380,243,367,257,359,279,390,274,384,306,395,308,390,329,407,337,407,374,414,390,433,407,403,423,413,433,457,444,447,458,459,506,484,519,488,528,469,540,480,556,463,558,497,576,497,588,481,592,494,604,487,621,463,616,457,626,487,655,436,675,384,668,392,694,394,705,355,703,327,707,319,681,315,698,283,689,266,701,247,688,247,674" shape="poly">
+          <area class="testMapArea" target="_blank" alt="tree" coords="227,661,140,709,132,698,103,690,131,668,76,650,83,637,56,629,33,644,-1,638,35,613,16,585,23,580,13,551,41,559,27,525,46,521,24,505,33,482,59,472,49,451,86,422,78,412,86,404,67,376,100,356,81,322,90,318,93,296,111,288,112,271,125,267,130,256,128,229,143,222,141,187,156,187,167,132,176,112,194,81,207,55,223,33,230,33,245,1,252,1,263,9,267,36,279,44,293,61,300,75,307,96,317,96,315,126,326,124,323,165,330,161,335,190,355,188,347,214,371,220,366,230,380,243,367,257,359,279,390,274,384,306,395,308,390,329,407,337,407,374,414,390,433,407,403,423,413,433,457,444,447,458,459,506,484,519,488,528,469,540,480,556,463,558,497,576,497,588,481,592,494,604,487,621,463,616,457,626,487,655,436,675,384,668,392,694,394,705,355,703,327,707,319,681,315,698,283,689,266,701,247,688,247,674" shape="poly">
       </map>
     </div>
     <div class="tree-toys col-3">
