+++ src/utils/utils.ts
@@ -36,8 +36,6 @@ const Utils = {
     img.width = 50;
     img.height = 50;
     img.draggable = true;
-    // img.id = `drag-${card.num}`;
-    // img.ondragstart = handleDragEvent;
     img.style.position = 'absolute';
     img.style.left = 'calc(50% - 25px)';
     img.style.top = 'calc(50% - 25px)';
